refactor: move ps/2 io operations to separate file
